pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lego-workflows"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
authors = [
    { name = "Tomas Perez Alvarez", email = "tomasperezalvarez@gmail.com" },
]
dependencies = ["result"]

[project.urls]
Documentation = "https://github.com/Tomperez98/lego-workflows#readme"
Issues = "https://github.com/Tomperez98/lego-workflows/issues"
Source = "https://github.com/Tomperez98/lego-workflows"

[project.optional-dependencies]
dev = ["pyrgo < 3", "types-toml", "pytest-asyncio", "pytest-cov"]

[tool.ruff.lint.isort]
known-first-party = ["lego_workflows"]
combine-as-imports = true
required-imports = ["from __future__ import annotations"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.flake8-tidy-imports.banned-api]
enum = { msg = "`enum` is banned: Python enums are not allowed. Use typing.Literal and typing.TypeAlias to build enums" }
mashumaro = { msg = "`mashumaro` is banned: We are using pydantic for data validation and serialization." }

[tool.ruff.lint.flake8-quotes]
inline-quotes = "double"

[tool.ruff]
line-length = 88
fix = false
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.lint]
ignore = [
    "ANN101",
    "D203",
    "D212",
    "COM812",
    "COM819",
    "D206",
    "E501",
    "ISC001",
    "Q000",
    "Q001",
    "Q002",
    "Q003",
]
select = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["INP001", "S101", "D"]
"scripts/*.py" = ["INP001"]

[tool.pytest.ini_options]
testpaths = ["tests"]
xfail_strict = true
addopts = [
    "--import-mode=importlib",
    "--strict-markers",
    "--cov=lego_workflows",
    "--cov-report=term-missing:skip-covered",
]
markers = ["integration: mark integration tests.", "unit: mark unittest."]
asyncio_mode = "auto"

[tool.coverage.report]
precision = 1
exclude_lines = [
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[tool.mypy]
follow_imports = "normal"
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
warn_unused_configs = true
disallow_subclassing_any = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
disallow_untyped_calls = true
disallow_untyped_defs = true
explicit_package_bases = true

[tool.pyrgo]
extra-paths = ["scripts"]
